Guards split into domain specific sub-modules
